pyproject.toml
[build-system]
requires = ["uv_build>=0.9.22,<0.10.0"]
build-backend = "uv_build"

[project]
name = "tero_project"
version = "0.0.1"
description = "M6 code structure test"
authors = [
  { name = "omez", email = "your@email.com" },
]
license = { file = "LICENSE" }
keywords = ["machine learning", "MLOps"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Programming Language :: Python :: 3",
]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "invoke==2.2.0",
    "fastapi==0.115.6",
    "gdown>=5.2.0",
    "torch==2.6.0",
    "typer==0.15.1",
    "uvicorn==0.34.0",
    "numpy<2.0",
    "matplotlib>=3.10.8",
    "scikit-learn>=1.7.2",
    "hydra-core>=1.3.2",
    "loguru>=0.7.3",
]
[project.scripts]


[dependency-groups]
dev = [
    "coverage==7.6.9",
    "mkdocs-material==9.4.6",
    "mkdocs==1.6.1",
    "pre-commit==4.1.0",
    "pytest==8.3.4",
    "ruff==0.1.3",
    "mkdocstrings-python==1.12.2",
]


[tool.uv.sources]
torch = [
    { index = "pytorch-cpu" },
]
torchvision = [
    { index = "pytorch-cpu" },
]

[[tool.uv.index]]
name = "pytorch-cpu"
url = "https://download.pytorch.org/whl/cpu"
explicit = true

[tool.ruff]
line-length = 120

[tool.coverage.run]
omit = ["tests/*"]
